Moved websiteSections to own file in _data, move TrustPositif fetching to getStaticProps
